requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cachelib==0.9.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
DateTime==5.1
dnacentersdk==2.6.1
fastjsonschema==2.16.3
Flask==2.2.3
Flask-Caching==2.0.2
fmcapi==20230322.0
frozenlist==1.3.3
future==0.18.3
idna==3.4
importlib-metadata==6.5.0
install==1.3.5
intersight==1.0.11.11558
ipaddress==1.0.23
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
meraki==1.32.1
multidict==6.0.4
pem==21.2.0
pycryptodome==3.17
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
requests-toolbelt==0.10.1
six==1.16.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
zope.interface==6.0
python-dotenv
